fix wrong inode size for rev#0 ext2fs
extfs.c assumes that there is always a valid inode_size field in the
superblock. But this is not true for ext2fs rev 0. Such ext2fs images are for
instance generated by genext2fs. Symptoms on ARM machines can be messages like:
"raise: Signal # 8 caught"
This fix checks for rev 0 and uses then 128 bytes as inode size.