Host tool to create an environment image
This is a host tool that creates a 4 KB environment
variable image. The environment image is by default
built with U8500 environment settings (file u8500.h is
included in the tool) but individual variables can be
removed, added and / or modified by specifying an
input configuration file containing such.

ST-Ericsson ID: WP272737

Change-Id: Ia5d79f72c9607ad25ed3d2e05795a7a616af9cf9
Signed-off-by: rickard evertsson <rickard.evertsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/5409
Reviewed-by: Michael BRANDT <michael.brandt@stericsson.com>