U8500: default env: MMC related fixes and changes
Enabled setexpr command to enable $filesize to $blocksize conversion in
scripts.

Changed occurrences of "mmc init" to "mmc rescan".
Changed paths in fat commands to absolute paths.

Added mmcboot alias to boot from external MMC (SD card).
The rootfs must be on the second partition. This allows to have a FAT FS
on the first partition; which is very common.

Change-Id: I17b89a43b4be7906f3f9e16e9ced9fa421a52d3a
Signed-off-by: Michael Brandt <michael.brandt@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/6964
Reviewed-by: Robert ROSENGREN <robert.rosengren@stericsson.com>
Reviewed-by: Kunal GOEL <kunal.goel@stericsson.com>