Prevent unnecessary build of u-boot
Even when the configuration had not really changed U-boot would be rebuilt
due to that the configuration files were updated unnecessarily. This change
prevents the configuration file to be updated as long as the config stays
the same, and this in turn prevents buildling U-boot. This applies to STE
Ux500 board configuration.

ST-Ericsson ID: AP280716
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Iacc39eee65d57578b6cbfcfc1dc8217b31a9c76f
Signed-off-by: Sebastian Rasmussen <sebastian.rasmussen@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/9228
Reviewed-by: Robert ROSENGREN <robert.rosengren@stericsson.com>
Reviewed-by: Michael BRANDT <michael.brandt@stericsson.com>
Tested-by: Michael BRANDT <michael.brandt@stericsson.com>