MeeGo: spec-file prepared for other variants
The u-boot spec-file is prepared for customized MeeGo products. The idea
is to add new spec files with the main spec as template, changing the
necessary configuration. From OBS perspective, a new linkpac should be
created with the same name as the spec file (hence triggering build of
the variant).

ST-Ericsson ID: AP280926
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ie51a3c2ecbcd5b71aefa988efa5bc3758079093f
Signed-off-by: Robert Rosengren <robert.rosengren@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/9338
Reviewed-by: Markus HELGESSON <markus.helgesson@stericsson.com>
Reviewed-by: Oscar EKBLADH <oscar.ekbladh@stericsson.com>
Reviewed-by: Olle TRANK <olle.trank@stericsson.com>
Reviewed-by: Robert MARKLUND <robert.marklund@stericsson.com>
Reviewed-by: Michael BRANDT <michael.brandt@stericsson.com>