MeeGo: Adding git version to boot log
The git version have been missing in the boot log when building in MeeGo,
since MeeGo is building in OBS and not in git source tree.

This patch adds a customized git2obs script that will both archive the
git and generate a correct version file. The version file will then
be used by the Makefile when building.

ST-Ericsson ID: 317452
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I22edc5d94b909307dd1539c4ff8645f771f12306
Signed-off-by: Robert Rosengren <robert.rosengren@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/11432
Reviewed-by: QATOOLS
Reviewed-by: Robert MARKLUND <robert.marklund@stericsson.com>
Reviewed-by: Michael BRANDT <michael.brandt@stericsson.com>