U8500: Add support for HREFP 2.0 V60 and later
The HREFP 2.0 V60 has no GPIO expander, therefore some GPIO lines
were moved around.
The preferred approach was to first check if there is a machid in the
environment. And if not try to program the GPIOE. If that fails, assume
it is a V60 or newer.
But for the splash screen, we need to know very early on which HW
version we are running. At that time no environment is available,
therefore it is unavoidable to do the I2C access (10 ms timeout), if
both HW versions shall be supported in one binary.

CAUTION: you need also a Linux kernel which supports MACH_TYPE_HREFV60.
Older U8500 kernels will not even issue a start-up message.

Depends-On: http://gerrit.lud.stericsson.com/gerrit/12335
ST-Ericsson ID: CR 272893

Change-Id: Id794df8c185dfbf67f7bce81deb49abf71129f7b
Signed-off-by: Michael Brandt <michael.brandt@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/12578
Reviewed-by: Torbjorn SVENSSON <torbjorn.x.svensson@stericsson.com>
Reviewed-by: Robert ROSENGREN <robert.rosengren@stericsson.com>
Reviewed-by: Markus HELGESSON <markus.helgesson@stericsson.com>
Reviewed-by: Bibek BASU <bibek.basu@stericsson.com>