Fix confusing "Sub TOC entry not found" messages
	MCDE:  Sub TOC entry not found
	toc_load_toc_entry: get_entry_info_toc failed
	mcde_display_image: no splash image found.
	ready

The splash code searches for the SPLASH entry. If it is not found in the
normal TOC, the TOC code in disk/part_toc.c tries the sub TOC entries.
The sub TOC code then issues the "Sub TOC entry not found" message.
It should be up to the caller to interpret it as error and to print an
appropriate error message.
Therefore this patch change it into a debug() message.

Change-Id: Ie2f566b0f7b8f4ff6a950ec05743b573c58626e7
Signed-off-by: Michael Brandt <michael.brandt@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/14828
Reviewed-by: QATOOLS
Reviewed-by: Joakim AXELSSON <joakim.axelsson@stericsson.com>
Reviewed-by: Markus HELGESSON <markus.helgesson@stericsson.com>